Move validators -> validator
Going to stick a validator class in there so it makes more sense it's singular
